backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-nanoid==2.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
